Use a named phoenix version and slim image
Two changes in one commit!

1. Use a named phoenix version. Before I was just pulling latest so that
I wouldn't have to update every time I used this. Now I changed to
naming the phoenix version as I'll be building the app from it so better
to have a named version.

2. Move to using the elixir slim image. I was using the full fat image
which meant each build was c.1Gb. These are now half the size (a quarter
if I didn't install node, which I don't actually use...)